Module Git.Fanout

Implementation of a fanout table (which orders value like the IDX file).

module type KEY = sig ... end

module Make : functor (Key : KEY) -> sig ... end

A Fanout table is a container to associate a key Key.t with a value. Internally, we order bindings by the first byte of the Key.t.