Module Fanout.Make

Parameters

module Key : KEY

Signature

type 'a t

The fanout table.

val make : unit -> 'a t

Make a new fanout table.

val bind : Key.t -> 'a -> 'a t -> unit

bind k v t adds the binding k and v to t.

val length : int -> 'a t -> int

length t returns the length of t.

val get : int -> 'a t -> (Key.t * 'a) list

get byte t gets a list of bindings where the key start with the byte byte.