Module Helper.MakeInflater

MakeInflater makes a module which respects the interface S.DECODER from an angstrom decoder and an inflate implementation.

This module use internallya MakeDecoder, so we have the same assertions

it's an extension of the previous decoder with an inflator. Then, this decoder decodes both the inflated flow and the serialized flow with fixed-size buffers.

Parameters

Z : Git__.S.INFLATE

A : Git__.S.DESC with type 'a t = 'a Angstrom.t

Signature

type t = A.e

type init = Z.window * Cstruct.t * Cstruct.t

type error = [

| Error.Decoder.t

| `Inflate of Z.error

]

type decoder

val pp_error : error Fmt.t

val to_result : Cstruct.t -> (t, error) Stdlib.result

val default : init -> decoder

val eval : decoder -> [ `Await of decoder | `End of Cstruct.t * t | `Error of Cstruct.t * error ]

val refill : Cstruct.t -> decoder -> (decoder, error) Stdlib.result

val finish : decoder -> decoder