Module Pack.Kind

This module represents the kind of a Git object - but only the kind, Kind.t does not have the Git value.

type t =

| Commit

| Tag

| Tree

| Blob
The type of kind.

val to_int : t -> int

to_int t returns an unique int value of the kind t. This value can be used to sort a list of kinds.

val to_bin : t -> int

to_bin t returns a binary code to serialize the kind t.

val pp : t Fmt.t

Pretty-printer of t.