Module Reference.P

type partial

Type of the left part of a reference.

type branch = string

Tyoe of the right part of a reference.

val (//) : partial -> string -> partial

Infix operator to compose string value with a partial reference.

val (/) : partial -> branch -> t

Infix operator to compose partial with branch and return a full-defined reference t.

val refs : partial

refs/

val heads : partial

refs/heads/

val remotes : partial

refs/remotes/

val origin : partial

refs/remotes/origin

val master : branch

*/master