Module Git.Tag

A Git Tag object.

A tag containing a reference pointing to another object, which can contain a message just like a Commit. It can also contain a (PGP) signature, in which case it is called a "signed tag object".

type kind =

| Blob

| Commit

| Tag

| Tree

type 'hash t

A Git Tag object. The tag object is very much like a Commit.t object - it contains a tagger, a date, a message, and a pointer. Generally, the tag points to a commit rather than a tree. It's like a branch reference, but it never moves - it always points to the same commit but gives it a friendlier name.

module type S = sig ... end

module Make (Hash : sig ... end) : S with type hash = Hash.t

Functor building an implementation of the tag structure. The functor returns a structure containing a type hash of digests and a type t of tags (structurally equal to t).