Module Git.Value

The Value module which represents the Git object.

module type S = sig ... end

Interface which describes the Git object.

module type RAW = sig ... end

Interface which describes raw operations. That means all serialization/unserialization to a Cstruct.t.

module Make : functor (Hash : sig ... end) -> functor (Inflate : sig ... end) -> functor (Deflate : sig ... end) -> S with module Hash := Hash and module Inflate := Inflate and module Deflate := Deflate and module Blob = Blob.Make(Hash) and module Commit = Commit.Make(Hash) and module Tree = Tree.Make(Hash) and module Tag = Tag.Make(Hash)

The functor to make the OCaml representation of the Git object by a specific hash implementation, an S.INFLATE implementation for the decompression and a S.DEFLATE implementation for the compression.

module Raw : functor (Hash : sig ... end) -> functor (Inflate : sig ... end) -> functor (Deflate : sig ... end) -> RAW with module Hash := Hash and module Inflate := Inflate and module Deflate := Deflate and module Blob = Blob.Make(Hash) and module Commit = Commit.Make(Hash) and module Tree = Tree.Make(Hash) and module Tag = Tag.Make(Hash) and module Value = Make(Hash)(Inflate)(Deflate) and type t = Make(Hash)(Inflate)(Deflate).t

The functor to make the OCaml representation of the Git object by a specific hash implementation, and S.INFLATE implementation for the decompression and a S.DEFLATE implementation for the compression.