Module Value.Tag

type t

A Git Tag object. The tag object is very much like a Commit.t object - it contains a tagger, a date, a message, and a pointer. Generally, the tag points to a commit rather than a tree. It's like a branch reference, but it never moves - it always points to the same commit but gives it a friendlier name.

type kind =

| Blob

| Commit

| Tag

| Tree

val make : Hash.t -> kind -> ?⁠tagger:User.t -> tag:string -> string -> t

make hash kind ?tagger ~tag descr makes a new tag to a Git object with the kind kind by the tagger with the name tag and the description descr.

This function does not check if the Git object pointed by the hash has really the kind kind - and obviously if the hash is valid.

module MakeMeta : functor (Meta : Encore.Meta.S) -> sig ... end

module A : sig ... end

module M : sig ... end

module D : sig ... end

module E : sig ... end

val digest : t -> Hash.t

val pp : t Fmt.t

val compare : t -> t -> int

val hash : t -> int

val equal : t -> t -> bool

module Set : Stdlib.Set.S with type elt = t

module Map : Stdlib.Map.S with type key = t

val length : t -> int64

length t returns the length of the tag object t.

val obj : t -> Hash.t

obj t returns the pointed hash of the Tag t.

val tag : t -> string

tag t returns the tag information of t.

val message : t -> string

val kind : t -> kind

val tagger : t -> User.t option