Module Loose.Make

Parameters

H : Git__.S.HASH

FS : Git__.S.FS

I : Git__.S.INFLATE

D : Git__.S.DEFLATE

Signature

module FS = FS

module Hash : Git__.S.HASH

module Deflate : Git__.S.DEFLATE

module Inflate : Git__.S.INFLATE

module Value : Git.Value.S with module Hash := Hash and module Deflate := Deflate and module Inflate := Inflate

include module type of Value

module Hash = Value.Hash

module Inflate = Value.Inflate

module Deflate = Value.Deflate

module Blob = Value.Blob

module Commit = Value.Commit

module Tree = Value.Tree

module Tag = Value.Tag

type t =

| Blob of Blob.t
The Blob.t OCaml value.

| Commit of Commit.t
The Commit.t OCaml value.

| Tree of Tree.t
The Tree.t OCaml value.

| Tag of Tag.t
The Tag.t OCaml value.

OCaml value which represents a Git object.

val blob : Blob.t -> t

val commit : Commit.t -> t

val tree : Tree.t -> t

val tag : Tag.t -> t

val kind : t -> [ `Commit | `Blob | `Tree | `Tag ]

kind o returns the kind of the Git object.

val pp_kind : [ `Commit | `Blob | `Tree | `Tag ] Fmt.t

pp_kind ppf kind is a human readable pretty-printer of kind.

module MakeMeta = Value.MakeMeta

module A = Value.A

module M = Value.M

module D = Value.D

module E = Value.E

include Git__.S.DIGEST with type t := t and type hash := Hash.t

type t

type hash

val digest : t -> hash

include Git__.S.BASE with type t := t

type t

val pp : t Fmt.t

val compare : t -> t -> int

val hash : t -> int

val equal : t -> t -> bool

module Set : Stdlib.Set.S with type Set.elt = t

module Map : Stdlib.Map.S with type Map.key = t

val length : t -> int64

type error = [

| Git.Error.Decoder.t

| Inflate.error Git.Error.Inf.t

| Deflate.error Git.Error.Def.t

| FS.error Git.Error.FS.t

]

type kind = [

| `Commit

| `Tree

| `Tag

| `Blob

]

val pp_error : error Fmt.t

val mem : fs:FS.t -> root:Fpath.t -> Hash.t -> bool Lwt.t

val read : fs:FS.t -> root:Fpath.t -> window:Inflate.window -> ztmp:Cstruct.t -> dtmp:Cstruct.t -> raw:Cstruct.t -> Hash.t -> (t, error) Stdlib.result Lwt.t

read ~fs ~root ~window ~ztmp ~dtmp ~raw hash tries to extract from a loose file localized in root / objects a Git object. This function reads (uses raw buffer to read) the file, inflates (uses ztmp and window to inflate) and decodes (uses dtmp to decode) Git object and make a Value.t.

val read_inflated : fs:FS.t -> root:Fpath.t -> window:Inflate.window -> ztmp:Cstruct.t -> dtmp:Cstruct.t -> raw:Cstruct.t -> Hash.t -> (kind * Cstruct.t, error) Stdlib.result Lwt.t

read_inflated ~fs ~root ~window ~ztmp ~dtmp ~raw hash tries to extract from a loose file localized in root / objects a Git object. This function reads (uses raw buffer to read) the file, inflates (uses ztmp and window to inflate) and decodes only Git header and body (uses dtmp to decode) Git object and make a fresh buffer.

Body is the serialization of the Git object, and, from kind (kind of Git object), client can call decoder associated to the kind returned (like Tree.Decoder for example).

val read_inflated_without_allocation : fs:FS.t -> root:Fpath.t -> window:Inflate.window -> ztmp:Cstruct.t -> dtmp:Cstruct.t -> raw:Cstruct.t -> result:Cstruct.t -> Hash.t -> (kind * Cstruct.t, error) Stdlib.result Lwt.t

read_inflated_without_allocation ~fs ~root ~window ~ztmp ~dtmp ~raw ~result hash is the same as read_inflated. However, instead to make a fresh buffer which will contain the Git object serialized, it uses result to store it. This function does not allocate any buffer.

val list : fs:FS.t -> root:Fpath.t -> Hash.t list Lwt.t

val size : fs:FS.t -> root:Fpath.t -> window:Inflate.window -> ztmp:Cstruct.t -> dtmp:Cstruct.t -> raw:Cstruct.t -> Hash.t -> (int64, error) Stdlib.result Lwt.t

val write : fs:FS.t -> root:Fpath.t -> temp_dir:Fpath.t -> etmp:Cstruct.t -> ?⁠level:int -> ztmp:Cstruct.t -> raw:Cstruct.t -> t -> (Hash.t * int, error) Stdlib.result Lwt.t

val write_deflated : fs:FS.t -> root:Fpath.t -> temp_dir:Fpath.t -> ?⁠level:int -> raw:Cstruct.t -> kind:kind -> Cstruct.t -> (Hash.t, error) Stdlib.result Lwt.t