Module Sync.Make

Creates a lower-level Sync functions fetch and push that are then overridden by backend-specific implementations such as Mem and Git_unix

Parameters

Digestif : Digestif.S

Pack : Smart_git.APPEND with type +'a fiber = 'a Lwt.t

Index : Smart_git.APPEND with type +'a fiber = 'a Lwt.t

Store : Git__.Minimal.S with type hash = Digestif.t

HTTP : Smart_git.HTTP

Signature

type hash = Digestif.t

type store = Store.t

type error = [

| `Exn of exn

| `Store of Store.error

| Mimic.error

]

val pp_error : error Fmt.t

val fetch : ?⁠push_stdout:(string -> unit) -> ?⁠push_stderr:(string -> unit) -> ?⁠threads:int -> ctx:Mimic.ctx -> Smart_git.Endpoint.t -> store -> ?⁠version:[> `V1 ] -> ?⁠capabilities:Smart.Capability.t list -> ?⁠deepen:[ `Depth of int | `Timestamp of int64 ] -> [ `All | `Some of (Git.Reference.t * Git.Reference.t) list | `None ] -> src:Pack.uid -> dst:Pack.uid -> idx:Index.uid -> create_idx_stream:(unit -> unit -> string option Lwt.t) -> create_pack_stream:(unit -> unit -> string option Lwt.t) -> Pack.t -> Index.t -> ((hash * (Git.Reference.t * hash) list) option, [> error ] as 'err) Stdlib.result Lwt.t

fetches remote references and saves them. Behavior of fetch when want is `All - fetches all remote references and saves them in store `Some src_dst_pairs - fetch src and save in dst `None - doesn't save anything

val push : ctx:Mimic.ctx -> Smart_git.Endpoint.t -> store -> ?⁠version:[> `V1 ] -> ?⁠capabilities:Smart.Capability.t list -> [ `Create of Git.Reference.t | `Delete of Git.Reference.t | `Update of Git.Reference.t * Git.Reference.t ] list -> (unit, [> error ] as 'err) Stdlib.result Lwt.t