Module Git__Helper

The Helper module should mnot be used by the client but only by maintainers of ocaml-git. It provides useful functions to interact with any I/O operations.

val ppe : name:string -> 'a Fmt.t -> 'a Fmt.t

ppe ~name pp make a new pretty-printer which shows your value like "(name %a)" pp value.

module Pair : sig ... end

module Option : sig ... end

module BaseIso : sig ... end

Helper for some convenience bijection elements.

module MakeDecoder : functor (A : Git__.S.DESC with type 'a t = 'a Angstrom.t) -> Git__.S.DECODER with type t = A.e and type init = Cstruct.t and type error = Git.Error.Decoder.t

MakeDecoder makes a module which respects the interface S.DECODER from an angstrom decoder.

module MakeInflater : functor (Z : Git__.S.INFLATE) -> functor (A : Git__.S.DESC with type 'a t = 'a Angstrom.t) -> Git__.S.DECODER with type t = A.e and type init = Z.window * Cstruct.t * Cstruct.t and type error = [ Git.Error.Decoder.t | `Inflate of Z.error ]

MakeInflater makes a module which respects the interface S.DECODER from an angstrom decoder and an inflate implementation.

module MakeEncoder : functor (M : Git__.S.DESC with type 'a t = 'a Encore.Encoder.t) -> Git__.S.ENCODER with type t = M.e and type init = Cstruct.t * M.e and type error = Git.Error.never

MakeEncoder makes a module which respects the interface S.ENCODER from a S.MINIENC.encoder. This module (instead MakeEncoder allocates one buffer from a size specified by the user. However this size must be a power of two.

module MakeDeflater : functor (Z : Git__.S.DEFLATE) -> functor (M : Git__.S.DESC with type 'a t = 'a Encore.Encoder.t) -> Git__.S.ENCODER with type t = M.e and type init = Cstruct.t * M.e * int * Cstruct.t and type error = [ `Deflate of Z.error ]

MakeDeflater makes a module which respects the interface S.ENCODER from a S.MINIENC.encoder. As MakeEncoder, this module allocates one buffer from a size specified by the user. However this size must be a power of two.

val digest : (module Git__.S.HASH with type HASH.t = 'hash) -> (module Git__.S.ENCODER with type error = Git.Error.never and type init = Cstruct.t * 't and type t = 't) -> etmp:Cstruct.t -> tmp:Cstruct.t -> kind:string -> length:('t -> int64) -> 't -> 'hash

fdigest (module Hash) (module Encoder) ?capacity ~tmp ~kind ~length value digests value with the Hash implementation and use the Encoder to stream on the Hash.digest function.

?capacity is the size of the internal buffer used to serialize your value (must be a power of two).

kind is the kind of the value (Commit, Blob, etc.) to make the Git header in top of the serialized value.

length is the function which calculate the weight of the value when it is serialized.

tmp is aan internal buffer used to store the stream of the encoder and used by Hash.digest.

module type ENCODER = sig ... end

module type DECODER = sig ... end

module Decoder : functor (D : DECODER) -> functor (FS : Git__.S.FS) -> sig ... end

module Encoder : functor (E : ENCODER) -> functor (FS : Git__.S.FS) -> sig ... end

module FS : functor (FS : Git__.S.FS) -> sig ... end