Module Value.Tag

module Hash : Git__.S.HASH

The Hash module used to make the implementation.

type t

A Git Tag object. The tag object is very much like a Commit.t object - it contains a tagger, a date, a message, and a pointer. Generally, the tag points to a commit rather than a tree. It's like a branch reference, but it never moves - it always points to the same commit but gives it a friendlier name.

type kind =

| Blob

| Commit

| Tag

| Tree

val make : Hash.t -> kind -> ?⁠tagger:Git.User.t -> tag:string -> string -> t

make hash kind ?tagger ~tag descr makes a new tag to a Git object with the kind kind by the tagger with the name tag and the description descr.

This function does not check if the Git object pointed by the hash has really the kind kind - and obviously if the hash is valid.

module MakeMeta : functor (Meta : Encore.Meta.S) -> sig ... end

module A : Git__.S.DESC with type 'a t = 'a Angstrom.t and type e = t

module M : Git__.S.DESC with type 'a t = 'a Encore.Encoder.t and type e = t

module D : Git__.S.DECODER with type t = t and type init = Cstruct.t and type error = Git.Error.Decoder.t

module E : Git__.S.ENCODER with type t = t and type init = Cstruct.t * t and type error = Git.Error.never

include Git__.S.DIGEST with type t := t and type hash := Hash.t

type t

type hash

val digest : t -> hash

include Git__.S.BASE with type t := t

type t

val pp : t Fmt.t

val compare : t -> t -> int

val hash : t -> int

val equal : t -> t -> bool

module Set : Stdlib.Set.S with type Set.elt = t

module Map : Stdlib.Map.S with type Map.key = t

val length : t -> int64

length t returns the length of the tag object t.

val obj : t -> Hash.t

obj t returns the pointed hash of the Tag t.

val tag : t -> string

tag t returns the tag information of t.

val message : t -> string

val kind : t -> kind

val tagger : t -> Git.User.t option