Module Decoder.Reference

module Hash : Git__.S.HASH

The Digest module used to make the module.

type nonrec t = Git.Reference.t

module P : sig ... end

val head : t

val master : t

val is_head : t -> bool

val of_string : string -> t

val to_string : t -> string

val of_path : Git.Path.t -> t

val to_path : t -> Git.Path.t

include Git__.S.BASE with type t := t

type t

val pp : t Fmt.t

val compare : t -> t -> int

val hash : t -> int

val equal : t -> t -> bool

module Set : Stdlib.Set.S with type Set.elt = t

module Map : Stdlib.Map.S with type Map.key = t

type head_contents =

| Hash of Hash.t
A pointer to an hash.

| Ref of t
A reference which one can point to an other reference or an hash.

The type of the value of a Git reference.

val pp_head_contents : head_contents Fmt.t

Pretty-printer of head_contents.

val equal_head_contents : head_contents -> head_contents -> bool

equal_head_contents a b implies a = Ref a' and b = Ref b' and Reference.equal a' b' = true or a = Hash a' and b = Hash b' and Hash.equal a' b'.

However, semantically Ref a' could be equal to Hash b' iff Hash b' is come from the reference a'. That means this function does not handle any indirection when it tests your values.

val compare_head_contents : head_contents -> head_contents -> int

module A : Git__.S.DESC with type 'a t = 'a Angstrom.t and type e = head_contents

module M : Git__.S.DESC with type 'a t = 'a Encore.Encoder.t and type e = head_contents

module D : Git__.S.DECODER with type t = head_contents and type init = Cstruct.t and type error = Git.Error.Decoder.t

module E : Git__.S.ENCODER with type t = head_contents and type init = Cstruct.t * head_contents and type error = Git.Error.never