Module Git__Value.Make

The functor to make the OCaml representation of the Git object by a specific hash implementation, an S.INFLATE implementation for the decompression and a S.DEFLATE implementation for the compression.

The constraints on git objects can cut the path to let the user to make the value module and keep the structural equality. If the Hash module is the same. That means:

```ocaml module V1 = Value.Make(SHA1)(C_inflate)(C_deflate) module V2 = Value.Make(SHA1)(OCaml_inflate)(OCaml_deflate) ```

Types V1.t and V2.t are equal.

Parameters

Hash : Git__.S.HASH

Inflate : Git__.S.INFLATE

Deflate : Git__.S.DEFLATE

Signature

module Hash : Git__.S.HASH

module Inflate : Git__.S.INFLATE

module Deflate : Git__.S.DEFLATE

module Blob = Git.Blob.Make(Hash)

module Commit = Git.Commit.Make(Hash)

module Tree = Git.Tree.Make(Hash)

module Tag = Git.Tag.Make(Hash)

type t =

| Blob of Blob.t
The Blob.t OCaml value.

| Commit of Commit.t
The Commit.t OCaml value.

| Tree of Tree.t
The Tree.t OCaml value.

| Tag of Tag.t
The Tag.t OCaml value.

OCaml value which represents a Git object.

val blob : Blob.t -> t

val commit : Commit.t -> t

val tree : Tree.t -> t

val tag : Tag.t -> t

val kind : t -> [ `Commit | `Blob | `Tree | `Tag ]

kind o returns the kind of the Git object.

val pp_kind : [ `Commit | `Blob | `Tree | `Tag ] Fmt.t

pp_kind ppf kind is a human readable pretty-printer of kind.

module MakeMeta : functor (Meta : Encore.Meta.S) -> sig ... end

module A : sig ... end

module M : Git__.S.DESC with type 'a t = 'a Encore.Encoder.t and type e = t

module D : Git__.S.DECODER with type t = t and type init = Inflate.window * Cstruct.t * Cstruct.t and type error = [ Git.Error.Decoder.t | `Inflate of Inflate.error ]

module E : Git__.S.ENCODER with type t = t and type init = Cstruct.t * t * int * Cstruct.t and type error = [ `Deflate of Deflate.error ]

include Git__.S.DIGEST with type t := t and type hash := Hash.t

type t

type hash

val digest : t -> hash

include Git__.S.BASE with type t := t

type t

val pp : t Fmt.t

val compare : t -> t -> int

val hash : t -> int

val equal : t -> t -> bool

module Set : Stdlib.Set.S with type Set.elt = t

module Map : Stdlib.Map.S with type Map.key = t

val length : t -> int64