Module Git.Commit

A Git Commit object.

A Git object which contains the information about a particular revision, such as parents, committed, author, date and Tree.t which corresponds to the top directory of the stored revision.

type 'hash t

A Git Commit object. Which specifies the top-level Tree.t for the snapshot of the project at a point; the author/committer information and the commit message.

module type S = sig ... end

module Make (Hash : sig ... end) : S with type hash = Hash.t

The functor to make the OCaml representation of the Git Commit object by a specific hash implementation.