Module Commit.Make

The functor to make the OCaml representation of the Git Commit object by a specific hash implementation.

Parameters

Hash : Git__.S.HASH

Signature

module Hash : Git__.S.HASH

type t

A Git Commit object. Which specifies the top-level Tree.t for the snapshot of the project at a point; the author/committer information and the commit message.

val make : tree:Hash.t -> author:User.t -> committer:User.t -> ?⁠parents:Hash.t list -> ?⁠extra:(string * string list) list -> string -> t

make ~author ~committer ?parents ~tree msg makes an OCaml value t. ?parents should be a non-empty list and corresponds to a list of hashes of commits. tree should be a hash of a Tree.t object.

This function does not write a new commit on the store and does not check the validity of parents and tree. By this way, this function never fails.

module MakeMeta : functor (Meta : Encore.Meta.S) -> sig ... end

module A : Git__.S.DESC with type 'a t = 'a Angstrom.t and type e = t

module M : Git__.S.DESC with type 'a t = 'a Encore.Encoder.t and type e = t

module D : Git__.S.DECODER with type t = t and type init = Cstruct.t and type error = Error.Decoder.t

module E : Git__.S.ENCODER with type t = t and type init = Cstruct.t * t and type error = Error.never

include Git__.S.DIGEST with type t := t and type hash := Hash.t

type t

type hash

val digest : t -> hash

include Git__.S.BASE with type t := t

type t

val pp : t Fmt.t

val compare : t -> t -> int

val hash : t -> int

val equal : t -> t -> bool

module Set : Stdlib.Set.S with type Set.elt = t

module Map : Stdlib.Map.S with type Map.key = t

val length : t -> int64

length t returns the length of the commit object t.

val parents : t -> Hash.t list

parents c returns all parents of the Git Commit object c.

val tree : t -> Hash.t

tree c returns the hash of top-level Tree.t of the Git Commit object c.

val committer : t -> User.t

committer c returns the committer of the commit c.

val author : t -> User.t

author c returns the author of the commit c.

val message : t -> string

message c returns the message of the commit c.

val extra : t -> (string * string list) list

val compare_by_date : t -> t -> int

compare_by_date a b compares the Git Commit object a and b by the date of the author. The compare function as the same behaviour.