Module type Object_graph.S

module Store : sig ... end

Store used as a graph.

module K : Graph.Sig.I with type V.t = Store.Hash.t

An imperative Graph of the store.

val keys : K.t -> Store.Hash.t list

keys graph returns all hashes recheables in the graph graph.

val of_keys : Store.t -> K.t Lwt.t

of_keys store makes a new graph from all values of a store.

val of_commits : Store.t -> K.t Lwt.t

of_commits store makes a new graph from all commits of a store.

val closure : ?⁠full:bool -> Store.t -> min:Store.Hash.Set.t -> max:Store.Hash.Set.t -> K.t Lwt.t

val pack : Store.t -> min:Store.Hash.Set.t -> max:Store.Hash.Set.t -> (Store.Hash.t * Store.Value.t) list Lwt.t

val to_dot : Store.t -> Stdlib.Format.formatter -> unit Lwt.t