Module Value.Raw

The functor to make the OCaml representation of the Git object by a specific hash implementation, and S.INFLATE implementation for the decompression and a S.DEFLATE implementation for the compression.

Parameters

Hash : Git__.S.HASH

Inflate : Git__.S.INFLATE

Deflate : Git__.S.DEFLATE

Signature

module Hash : Git__.S.HASH

module Inflate : Git__.S.INFLATE

module Deflate : Git__.S.DEFLATE

module Value = Make(Hash)(Inflate)(Deflate)

include module type of Value

module Blob : sig ... end

module Commit : sig ... end

module Tree : sig ... end

module Tag : sig ... end

type t =

| Blob of Blob.t
The Blob.t OCaml value.

| Commit of Commit.t
The Commit.t OCaml value.

| Tree of Tree.t
The Tree.t OCaml value.

| Tag of Tag.t
The Tag.t OCaml value.

OCaml value which represents a Git object.

val blob : Blob.t -> t

val commit : Commit.t -> t

val tree : Tree.t -> t

val tag : Tag.t -> t

val kind : t -> [ `Blob | `Commit | `Tag | `Tree ]

kind o returns the kind of the Git object.

val pp_kind : [ `Blob | `Commit | `Tag | `Tree ] Fmt.t

pp_kind ppf kind is a human readable pretty-printer of kind.

module MakeMeta : functor (Meta : Encore.Meta.S) -> sig ... end

module A : sig ... end

module M : sig ... end

module D : sig ... end

module E : sig ... end

val digest : t -> Hash.t

val pp : t Fmt.t

val compare : t -> t -> int

val hash : t -> int

val equal : t -> t -> bool

module Set : sig ... end

module Map : sig ... end

val length : t -> int64

module EncoderRaw : Git__.S.ENCODER with type t = t and type init = Cstruct.t * t and type error = Error.never

module DecoderRaw : Git__.S.DECODER with type t = t and type init = Cstruct.t and type error = Error.Decoder.t

module EncoderWithoutHeader : Git__.S.ENCODER with type t = t and type init = Cstruct.t * t and type error = Error.never

val to_deflated_raw : raw:Cstruct.t -> etmp:Cstruct.t -> ?⁠level:int -> ztmp:Cstruct.t -> t -> (string, E.error) Stdlib.result

to_deflated_raw ?capacity ?level ~ztmp value serializes and deflates the value value. capacity is the memory consumption of the encoder in bytes (default to 0x100), level is the level of the compression (default to 4) and ztmp is an internal buffer used to store the serialized value before the deflation.

All error from the Deflate module is relayed to the `Deflate error value.

val to_raw : raw:Cstruct.t -> etmp:Cstruct.t -> t -> (string, EncoderRaw.error) Stdlib.result

to_raw ?capacity value serializes the value value. capacity is the memory consumption of the encoder in bytes (default to 0x100).

This function can not returns an EE.error (see EE).

val to_raw_without_header : raw:Cstruct.t -> etmp:Cstruct.t -> t -> (string, EncoderWithoutHeader.error) Stdlib.result

val of_raw : kind:[ `Commit | `Blob | `Tree | `Tag ] -> Cstruct.t -> (t, Error.Decoder.t) Stdlib.result

of_raw ~kind inflated makes a Git object as an OCaml value t. This decoder does not expect an header to recognize which kind of Git object is it. That means the inflated raw should not contain kind size\000 at the beginning (in this case, you should use of_raw_with_header.

val of_raw_with_header : Cstruct.t -> (t, DecoderRaw.error) Stdlib.result

of_raw_with_header inflated makes a Git object as an OCaml value t. This decoder expects an header to choose which Git object it is.